refactor(audio): move the audio console verbs to the module that owns them
playsample and the audio subcommand tree were 217 lines inside CONSOLE_CMD.CPP,
which is the generic console surface. CODESTYLE's rule is that a surface offers a
registration hook and the feature fills it in, which is how the camera's camtrace
already works; audio was the counter-example sitting next to it.

They move to AUDIO_CONSOLE.CPP behind Audio_RegisterCommands, called from
Console_RegisterAll beside FollowCam_RegisterCommands. CONSOLE_CMD.CPP goes from
2494 lines to 2269 and no longer names an audio function.

A separate file rather than more of AMBIANCE.CPP, because this is bindings and
not mechanism: it parses argv, prints, and calls the module. SOURCES/CONSOLE/
already draws that line between its core and CONSOLE_CMD.CPP, and the same one
makes the mechanism side reachable from a host test later.

A pure move: the command bodies are unchanged, and the two registration calls
keep their help text verbatim. Verified by running the verbs, and by checking
`help` still lists both.
