+++ SOURCES/CMakeLists.txt
@@ -17,7 +17,7 @@ add_custom_command(
 )
 
 # General sources
-set(SOURCES_FILES AMBIANCE.CPP DISTRIB.CPP AUDIO_BALANCE.CPP ANIMTEX.CPP BEZIER.CPP BUILD_INFO.CPP BUGGY.CPP CHEATCOD.CPP CLI_ARGS.CPP COMPORTE.CPP CONFIG.CPP CONTROL.CPP CONTROL_PROTO.CPP CONTROL_SERVER.CPP CREDITS.CPP CREDITS_PARSE.CPP
+set(SOURCES_FILES AMBIANCE.CPP DISTRIB.CPP AUDIO_BALANCE.CPP AUDIO_CONSOLE.CPP ANIMTEX.CPP BEZIER.CPP BUILD_INFO.CPP BUGGY.CPP CHEATCOD.CPP CLI_ARGS.CPP COMPORTE.CPP CONFIG.CPP CONTROL.CPP CONTROL_PROTO.CPP CONTROL_SERVER.CPP CREDITS.CPP CREDITS_PARSE.CPP
         CONFIG_FILE.CPP DART.CPP DEMO_SEED.CPP DIRECTORIES.CPP DISKFUNC.CPP EMBEDDED_CFG_WRITE.CPP EXIT_SCREEN.CPP EXTFUNC.CPP EXTRA.CPP FICHE.CPP FLOW.CPP FOLLOWCAM.CPP GAMEMENU.CPP GAMEMENU_MOUSE.CPP MENU_KEYNAV.CPP MENU_LABELS.CPP GERELIFE.CPP
         GERETRAK.CPP GLOBAL.CPP GRILLE.CPP HOLOBODY.CPP HOLOGLOB.CPP HOLOPLAN.CPP IMPACT.CPP INCRUST.CPP INITADEL.C
         INPUT.CPP INTEXT.CPP INVENT.CPP JOYSTICK.CPP LZSS.CPP MEM.CPP MESSAGE.CPP MUSIC.CPP OBJECT.CPP PATCH.CPP IMAGE_LOAD.CPP PERFTRACE.CPP PERSO.CPP
